wordpress and mariadb script

+++ srcs/requirements/wordpress/tools/script.sh
@@ -1,28 +1,37 @@
 #!/bin/bash
 
+GREEN='\033[0;32m'
+RED='\033[0;31m'
+NC='\033[0m'
+
 WORDPRESS_URL="https://wordpress.org/latest.tar.gz"
 WORDPRESS_DIR="/var/www/html"
 WP_CONFIG_SOURCE="/wp-config.php"
 
-echo "Descargando WordPress..."
-curl -sSL "$WORDPRESS_URL" -o /tmp/wordpress.tar.gz
+if [ ! -f "$WORDPRESS_DIR/index.php" ]; then
+    echo "${GREEN}Descargando WordPress...${NC}"
+    curl -sSL "$WORDPRESS_URL" -o /tmp/wordpress.tar.gz
 
-echo "Descomprimiendo WordPress..."
-tar -xzf /tmp/wordpress.tar.gz -C /tmp
+    echo "${GREEN}Descomprimiendo WordPress...${NC}"
+    tar -xzf /tmp/wordpress.tar.gz -C /tmp
 
-echo "Copiando archivos de WordPress a $WORDPRESS_DIR..."
-cp -r /tmp/wordpress/* "$WORDPRESS_DIR/"
+    echo "${GREEN}Copiando archivos de WordPress a $WORDPRESS_DIR...${NC}"
+    cp -r /tmp/wordpress/* "$WORDPRESS_DIR/"
 
-echo "Limpiando archivos temporales..."
-rm -rf /tmp/wordpress /tmp/wordpress.tar.gz
+    echo "${GREEN}Limpiando archivos temporales...${NC}"
+    rm -rf /tmp/wordpress /tmp/wordpress.tar.gz
+    rm -rf "$WORDPRESS_DIR/wp-config-sample.php"
 
-mkdir -p /var/www/html
+    mkdir -p /var/www/html
 
-echo "Copiando wp-config.php ..."
-cp "$WP_CONFIG_SOURCE" "$WORDPRESS_DIR/wp-config.php"
+    echo "${GREEN}Copiando wp-config.php ...${NC}"
+    cp "$WP_CONFIG_SOURCE" "$WORDPRESS_DIR/wp-config.php"
 
-echo "Ajustando permisos..."
-chown -R www-data:www-data "$WORDPRESS_DIR"
-chmod -R 755 "$WORDPRESS_DIR"
+    echo "${GREEN}Ajustando permisos...${NC}"
+    chown -R www-data:www-data "$WORDPRESS_DIR"
+    chmod -R 755 "$WORDPRESS_DIR"
 
-echo "¡WordPress instalado y configurado correctamente!"
+    echo "${GREEN}¡WordPress instalado y configurado correctamente!${NC}"
+else
+    echo "${RED}WordPress ya está instalado. No es necesario ejecutar el script nuevamente.${NC}"
+fi
